requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
huggingface-hub>=0.15.0
wandb>=0.15.0
tensorboard>=2.12.0
peft>=0.4.0
sentencepiece>=0.1.99
pyyaml>=6.0
tqdm>=4.65.0